refactor: Move interfaces (#336)
* feat: move interfaces

* fix: References in application project

* fix: References in web project

* fix: references in tests projects

* fix: view imports

* fix: view import
