+++ src/Dfe.PlanTech.Application/Submissions/Commands/SubmitAnswerCommand.cs
@@ -1,8 +1,8 @@
 using Dfe.PlanTech.Application.Persistence.Interfaces;
-using Dfe.PlanTech.Application.Submissions.Interfaces;
-using Dfe.PlanTech.Application.Users.Interfaces;
 using Dfe.PlanTech.Domain.Questionnaire.Models;
 using Dfe.PlanTech.Domain.Responses.Models;
+using Dfe.PlanTech.Domain.Submissions.Interfaces;
+using Dfe.PlanTech.Domain.Users.Interfaces;
 using Microsoft.Data.SqlClient;
 using System.Data;
 using System.Security.Authentication;
