+++ src/Dfe.PlanTech.Application/Users/Commands/CreateUserCommand.cs
@@ -1,4 +1,5 @@
 using Dfe.PlanTech.Application.Persistence.Interfaces;
+using Dfe.PlanTech.Domain.Users.Interfaces;
 using Dfe.PlanTech.Domain.Users.Models;
 
 namespace Dfe.PlanTech.Application.Users.Commands;
