+++ src/Dfe.PlanTech.Application/Users/Commands/RecordUserSigninCommand.cs
@@ -1,7 +1,7 @@
 ﻿using Dfe.PlanTech.Application.Persistence.Interfaces;
-using Dfe.PlanTech.Application.Users.Interfaces;
 using Dfe.PlanTech.Domain.Establishments.Models;
 using Dfe.PlanTech.Domain.SignIns.Models;
+using Dfe.PlanTech.Domain.Users.Interfaces;
 using Dfe.PlanTech.Domain.Users.Models;
 namespace Dfe.PlanTech.Application.Users.Commands;
