+++ src/Dfe.PlanTech.Application/Users/Queries/GetUserIdQuery.cs
@@ -1,5 +1,5 @@
 using Dfe.PlanTech.Application.Persistence.Interfaces;
-using Dfe.PlanTech.Application.Users.Interfaces;
+using Dfe.PlanTech.Domain.Users.Interfaces;
 
 namespace Dfe.PlanTech.Application.Users.Queries;
