+++ src/Dfe.PlanTech.Domain/Submissions/Interfaces/ISubmitAnswerCommand.cs
@@ -1,6 +1,6 @@
 using Dfe.PlanTech.Domain.Questionnaire.Models;
 
-namespace Dfe.PlanTech.Application.Submissions.Interfaces;
+namespace Dfe.PlanTech.Domain.Submissions.Interfaces;
 
 public interface ISubmitAnswerCommand
 {
