+++ src/Dfe.PlanTech.Web/Views/Shared/_CookieBanner.cshtml
@@ -1,5 +1,5 @@
 ﻿@using Microsoft.AspNetCore.Http.Features
-@inject ICookieService cookieService
+@inject Dfe.PlanTech.Domain.Cookie.Interfaces.ICookieService cookieService
 @{
     var cookie = cookieService.GetCookie();  
 }
