+++ src/Dfe.PlanTech.Web/Views/_ViewImports.cshtml
@@ -1,7 +1,5 @@
 ﻿@using Dfe.PlanTech.Web
 @using Dfe.PlanTech.Web.Models
-@using Dfe.PlanTech.Application.Submissions.Queries
-@using Dfe.PlanTech.Application.Cookie.Interfaces
 @addTagHelper *, GovUk.Frontend.AspNetCore
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
 @addTagHelper *, Dfe.PlanTech.Web
