+++ tests/Dfe.PlanTech.Application.UnitTests/Responses/Commands/ProcessCheckAnswerDtoCommandTests.cs
@@ -1,7 +1,7 @@
 using Dfe.PlanTech.Application.Responses.Commands;
-using Dfe.PlanTech.Application.Responses.Interface;
 using Dfe.PlanTech.Domain.Content.Models;
 using Dfe.PlanTech.Domain.Questionnaire.Models;
+using Dfe.PlanTech.Domain.Responses.Interface;
 using NSubstitute;
 
 namespace Dfe.PlanTech.Application.UnitTests.Responses.Commands;
