+++ tests/Dfe.PlanTech.Application.UnitTests/Submissions/Commands/SubmitAnswerCommandTests.cs
@@ -1,8 +1,8 @@
 using Dfe.PlanTech.Application.Persistence.Interfaces;
 using Dfe.PlanTech.Application.Submissions.Commands;
-using Dfe.PlanTech.Application.Submissions.Interfaces;
-using Dfe.PlanTech.Application.Users.Interfaces;
 using Dfe.PlanTech.Domain.Questionnaire.Models;
+using Dfe.PlanTech.Domain.Submissions.Interfaces;
+using Dfe.PlanTech.Domain.Users.Interfaces;
 using Microsoft.Data.SqlClient;
 using NSubstitute;
 using System.Security.Authentication;
