+++ tests/Dfe.PlanTech.Application.UnitTests/Users/Commands/RecordUserSigninCommandTests.cs
@@ -1,7 +1,7 @@
 ﻿using Dfe.PlanTech.Application.Persistence.Interfaces;
 using Dfe.PlanTech.Application.Users.Commands;
-using Dfe.PlanTech.Application.Users.Interfaces;
 using Dfe.PlanTech.Domain.SignIns.Models;
+using Dfe.PlanTech.Domain.Users.Interfaces;
 using Dfe.PlanTech.Domain.Users.Models;
 using NSubstitute;
 using NSubstitute.ReturnsExtensions;
