+++ tests/Dfe.PlanTech.Infrastructure.SignIn.UnitTests/DfeOpenIdConnectEventsTests.cs
@@ -1,7 +1,7 @@
 using Dfe.PlanTech.Application.SignIns.Interfaces;
-using Dfe.PlanTech.Application.Users.Interfaces;
 using Dfe.PlanTech.Domain.SignIns.Enums;
 using Dfe.PlanTech.Domain.SignIns.Models;
+using Dfe.PlanTech.Domain.Users.Interfaces;
 using Dfe.PlanTech.Domain.Users.Models;
 using Dfe.PlanTech.Infrastructure.SignIns.ConnectEvents;
 using Microsoft.AspNetCore.Authentication;
